+++ src/main/kotlin/graphql/dataTransferObjects/StarredMessage.kt
@@ -2,19 +2,15 @@
 
 package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
-import com.neelkamath.omniChatBackend.db.tables.MessageStatuses
+import com.neelkamath.omniChatBackend.db.ForwardPagination
 import com.neelkamath.omniChatBackend.db.tables.Messages
 import com.neelkamath.omniChatBackend.graphql.routing.MessageState
+import graphql.schema.DataFetchingEnvironment
 import java.time.LocalDateTime
 
 sealed interface StarredMessage {
     /** The [Messages.id]. */
     val id: Int
-
-    /** The ID of the chat which this message this belongs to. */
-    val chatId: Lazy<Int>
-
-    fun getChatId(): Int = chatId.value
 
     fun getMessageId(): Int = id
 
@@ -24,10 +20,17 @@ sealed interface StarredMessage {
 
     fun getSent(): LocalDateTime = Messages.readSent(id)
 
-    fun getStatuses(): List<MessageDateTimeStatus> =
-        MessageStatuses.readIdList(id).map(::MessageDateTimeStatus)
+    fun getStatuses(env: DataFetchingEnvironment): MessageDateTimeStatusConnection {
+        val pagination = ForwardPagination(env.getArgument("first"), env.getArgument("after"))
+        return MessageDateTimeStatusConnection(id, pagination)
+    }
 
     fun getContext(): MessageContext = MessageContext(id)
 
     fun getIsForwarded(): Boolean = Messages.isForwarded(id)
+
+    /** The ID of the chat which this message this belongs to. */
+    val chatId: Lazy<Int>
+
+    fun getChatId(): Int = chatId.value
 }