+++ src/main/kotlin/graphql/dataTransferObjects/UpdatedMessage.kt
@@ -2,7 +2,7 @@
 
 package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
-import com.neelkamath.omniChatBackend.db.tables.MessageStatuses
+import com.neelkamath.omniChatBackend.db.ForwardPagination
 import com.neelkamath.omniChatBackend.db.tables.Messages
 import com.neelkamath.omniChatBackend.db.tables.Stargazers
 import com.neelkamath.omniChatBackend.graphql.routing.MessageState
@@ -16,7 +16,10 @@ class UpdatedMessage(private val messageId: Int) : MessagesSubscription, ChatMes
 
     fun getState(): MessageState = Messages.readState(messageId)
 
-    fun getStatuses(): List<MessageDateTimeStatus> = MessageStatuses.readIdList(messageId).map(::MessageDateTimeStatus)
+    fun getStatuses(env: DataFetchingEnvironment): MessageDateTimeStatusConnection {
+        val pagination = ForwardPagination(env.getArgument("first"), env.getArgument("last"))
+        return MessageDateTimeStatusConnection(messageId, pagination)
+    }
 
     fun getHasStar(env: DataFetchingEnvironment): Boolean = Stargazers.hasStar(env.userId!!, messageId)
 }